Add Erwerbsminderungsrente (#577)
Implementation of pension for reduced earning capacity
(Erwerbsminderungsrente) as addition to the pension system already
existing in gettsim.

The special feature of the pension for reduced earning capacity are the
additional earning points. Average earning points per year are
calculated for every individual and attributed for the time between
retirement and an allocation age limit (Zurechnungszeitsgrenze). There
are also special rules for (Zugangsfaktor) and (Rentenfaktor) both
relevant for the calculation of the total pension claim.

Two new input variables are introduced.

- `voll_erwerbsgemind`: Unable to provide more than 3 hours of market labor per day.
- `teilw_erwerbsgemind`: Able to provide at least 3 but no more than 6 hours of market labor per day.

---------

Co-authored-by: Stefan Klotz <kl.stefan@gmx.de>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Annica Gehlen <39128048+amageh@users.noreply.github.com>
Co-authored-by: Hans-Martin von Gaudecker <hmgaudecker@gmail.com>