Fix time conversion after rounding (#715)
### What problem do you want to solve?

Closes #712

Previously, deriving a time-converted function from a rounded function
lead to the following error:

```txt
KeyError: Rounding specifications for function test_func_y are expected in the parameter dictionary 
at ['params_key_test']['rounding']['test_func_y']. These nested keys do not exist. 
If this function should not be rounded, remove the respective decorator.
```

This PR fixes this.

### Todo

- [X] Pick an appropriate title.
- [X] Put `Closes #XXXX` in the first PR comment to auto-close the
relevant issue once
the PR is accepted. This is not applicable if there is no corresponding
issue.
- [ ] ~Document PR in CHANGES.md.~ (Omitted due to small size of PR)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>