Adjust statutory retirement ages (#717)
Closes #665, #664
Closes #491 partially
Related to #663 

### What problem do you want to solve?

This PR solves two problems (as discussed in #665):
1. Statutory retirement ages (SRAs) are wrong in the parameter files.
This is a relict of the missing `dates_active` feature. Back then, we
dealt with abolished pathways into retirement by setting the relevant
SRA of the abolished pathway to the SRA of a pathway that continued to
exist (e.g. setting the full retirement age for Altersrente for Frauen
to the Regelaltersgrenze after the 1952 birth cohort). This logic fails
once we want to depict changes of the policy environment over time.
2. SRAs are currently implemented via piecewise linear functions. We
started to replace this with explicit birth-cohort-age dicts in #645

Along these lines, this PR changes the test files that currently only
test the pension functions for policy year 2010. Due to point 1, there
are probably a lot of tests that expect wrong SRAs.

### Todo

- [x] Parameters + Functions of Regelaltersgrenze / NRA
- [x] Parameters + Functions of Altersrente f. Frauen (abolished since
birth cohort 1952)
- [x] Parameters + Functions of Altersrente f. langj. Versicherte
- [x] Parameters + Functionf of Altersrente f. bes. langj. Versicherte
(since 2012)

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>