Revert "refactor: class for derived functions"
This reverts commit 4afa60bd0963a9c5962c9a35562dcd05c4adba02.

The current PolicyEnvironment currently expects only PolicyFunctions. We should not widen this to general GETTSIMFunctions, since the user should not create derived functions manually. As we also read from the functions list/dictionary, subclassing PolicyEnvironment and overriding the field is incorrect (dict is invariant). Thus, we would need to duplicate much of the functionality of PolicyEnvironment, just to handle derived classes.

It seems more elegant to add fields to PolicyFunction or let a future DerivedFunction extend PolicyFunction.