2025 social security and pension parameters (#809)
### What problem do you want to solve?
addresses #806 by adding changes to social security contributions and
pensions .

To be discussed:
There is no differentiation anymore between East and West Germany with
regard to social security and pensions. The easy way (as done here) is
to set the `east` and `west` parameter identical henceforth.
Alternatively, we could create new functions which ignore the
`wohnort_ost` variable.


### Todo

- [x] Pick an appropriate title.
- [x] Put `Closes #XXXX` in the first PR comment to auto-close the
relevant issue once
the PR is accepted. This is not applicable if there is no corresponding
issue.
- [x] Document PR in CHANGES.md.

---------

Co-authored-by: Eric Sommer <eric-sommer@github.com>