Steuerfortentwicklungsgesetz 2024 (#807)
### What problem do you want to solve?

addresses #806 

Implement the parameter changes from 2. Jahressteuergesetz 2024. (see
[here](https://bundestagszusammenfasser.de/details?docid=667). It is not
yet officially published, there is hence no official reference yet.

The Kinderzuschlag issue is also not clear to me. The parameter
`kindersofortzuschl` which has been inactive in gettsim since 2023, is
going to be changed by this law.

### Todo

- [x] Pick an appropriate title.
- [x] Put `Closes #XXXX` in the first PR comment to auto-close the
relevant issue once
the PR is accepted. This is not applicable if there is no corresponding
issue.
- [x] Document PR in CHANGES.md.

---------

Co-authored-by: Eric Sommer <eric-sommer@github.com>
Co-authored-by: Marvin Immesberger <immesberger@posteo.de>
Co-authored-by: Maximilian Blömer <bloemer@ifo.de>
Co-authored-by: Marvin Immesberger <74215010+MImmesberger@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>