Change behaviour of namespaces so they are at the module level. (#818)
The way we implemented the loading of namespaces in #780 does not quite
work.

We want to have them at the directory level to balance use of namespaces
and reducing the amount of qualified names.

Additionally, we had to change the order of the upsert operations in `combine_policy_functions_and_derived_functions`. Doesn't affect the happy path, but in case of conflicts the previous behaviour did not make sense.

---------

Co-authored-by: Marvin Immesberger <immesberger@posteo.de>